+++ cmd/mountv1_test.go
@@ -36,7 +36,6 @@ package cmd
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"testing"
@@ -115,8 +114,8 @@ func generateFileName() string {
 func (suite *generateConfigTestSuite) TestConfigFileInvalid() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName myOtherAccountName")
@@ -128,8 +127,8 @@ func (suite *generateConfigTestSuite) TestConfigFileInvalid() {
 func (suite *generateConfigTestSuite) TestConfigFileKey() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\naccountKey myAccountKey\nauthType Key\ncontainerName myContainerName\n")
@@ -154,8 +153,8 @@ func (suite *generateConfigTestSuite) TestConfigFileKey() {
 func (suite *generateConfigTestSuite) TestConfigFileSas() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nsasToken mySasToken\nauthType SAS\ncontainerName myContainerName\n")
@@ -180,8 +179,8 @@ func (suite *generateConfigTestSuite) TestConfigFileSas() {
 func (suite *generateConfigTestSuite) TestConfigFileSPN() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nservicePrincipalClientId clientId\nservicePrincipalTenantId tenantId\nservicePrincipalClientSecret clientSecret\naadEndpoint aadEndpoint\nauthType SPN\ncontainerName myContainerName\n")
@@ -208,8 +207,8 @@ func (suite *generateConfigTestSuite) TestConfigFileSPN() {
 func (suite *generateConfigTestSuite) TestConfigFileMSI() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nidentityClientId clientId\nidentityObjectId objectId\nidentityResourceId resourceId\nauthType MSI\ncontainerName myContainerName\n")
@@ -236,8 +235,8 @@ func (suite *generateConfigTestSuite) TestConfigFileMSI() {
 func (suite *generateConfigTestSuite) TestConfigFileProxy() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nhttpProxy httpProxy\nhttpsProxy httpsProxy\n")
@@ -259,8 +258,8 @@ func (suite *generateConfigTestSuite) TestConfigFileProxy() {
 func (suite *generateConfigTestSuite) TestConfigFileBlobEndpoint() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nblobEndpoint blobEndpoint\n")
@@ -281,8 +280,8 @@ func (suite *generateConfigTestSuite) TestConfigFileBlobEndpoint() {
 func (suite *generateConfigTestSuite) TestConfigFileAccountType() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\naccountType adls\n")
@@ -304,8 +303,8 @@ func (suite *generateConfigTestSuite) TestConfigFileAccountType() {
 func (suite *generateConfigTestSuite) TestConfigFileAuthMode() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nauthType Key\n")
@@ -326,8 +325,8 @@ func (suite *generateConfigTestSuite) TestConfigFileAuthMode() {
 func (suite *generateConfigTestSuite) TestConfigFileLogLevel() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nlogLevel LOG_ERROR\n")
@@ -348,8 +347,8 @@ func (suite *generateConfigTestSuite) TestConfigFileLogLevel() {
 func (suite *generateConfigTestSuite) TestConfigFileIgnoreCommentsNewLine() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nlogLevel LOG_ERROR\n# accountName myAccountName\n")
@@ -370,8 +369,8 @@ func (suite *generateConfigTestSuite) TestConfigFileIgnoreCommentsNewLine() {
 func (suite *generateConfigTestSuite) TestConfigFileIgnoreCommentsSameLine() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nlogLevel LOG_ERROR #LOG_DEBUG\n")
@@ -392,8 +391,8 @@ func (suite *generateConfigTestSuite) TestConfigFileIgnoreCommentsSameLine() {
 func (suite *generateConfigTestSuite) TestConfigFileCaCertFileError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\ncaCertFile caCertFile\n")
@@ -405,8 +404,8 @@ func (suite *generateConfigTestSuite) TestConfigFileCaCertFileError() {
 func (suite *generateConfigTestSuite) TestConfigFileDnsTypeError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\ndnsType dnsType\n")
@@ -418,8 +417,8 @@ func (suite *generateConfigTestSuite) TestConfigFileDnsTypeError() {
 func (suite *generateConfigTestSuite) TestConfigCLILogLevel() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v1ConfigFile.Name())
 	defer os.Remove(v2ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\nlogLevel LOG_ERROR\n")
@@ -440,11 +439,11 @@ func (suite *generateConfigTestSuite) TestConfigCLILogLevel() {
 func (suite *generateConfigTestSuite) TestCLIParamLogging() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
 
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -467,10 +466,10 @@ func (suite *generateConfigTestSuite) TestCLIParamFileCache() {
 	defer suite.cleanupTest()
 	name := generateFileName()
 
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -504,10 +503,10 @@ func (suite *generateConfigTestSuite) TestCLIParamFileCache() {
 func (suite *generateConfigTestSuite) TestAddStreamAndFileCache() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -532,10 +531,10 @@ func (suite *generateConfigTestSuite) TestAddStreamAndFileCache() {
 func (suite *generateConfigTestSuite) TestComponentCorrectOrder() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -560,10 +559,10 @@ func (suite *generateConfigTestSuite) TestComponentCorrectOrder() {
 func (suite *generateConfigTestSuite) TestCLIParamFileCacheUploadModifiedOnlyError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -576,10 +575,10 @@ func (suite *generateConfigTestSuite) TestCLIParamFileCacheUploadModifiedOnlyErr
 func (suite *generateConfigTestSuite) TestCLIParamFileCachePollTimeoutError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -592,10 +591,10 @@ func (suite *generateConfigTestSuite) TestCLIParamFileCachePollTimeoutError() {
 func (suite *generateConfigTestSuite) TestCLIParamStreaming() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -622,10 +621,10 @@ func (suite *generateConfigTestSuite) TestCLIParamStreaming() {
 func (suite *generateConfigTestSuite) TestCLIParamAttrCache() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -650,10 +649,10 @@ func (suite *generateConfigTestSuite) TestCLIParamAttrCache() {
 func (suite *generateConfigTestSuite) TestCLIParamStorage() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -693,10 +692,10 @@ func (suite *generateConfigTestSuite) TestCLIParamStorage() {
 func (suite *generateConfigTestSuite) TestCLIParamStorageCaCertFileError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -709,10 +708,10 @@ func (suite *generateConfigTestSuite) TestCLIParamStorageCaCertFileError() {
 func (suite *generateConfigTestSuite) TestCLIParamStorageContentTypeError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -725,10 +724,10 @@ func (suite *generateConfigTestSuite) TestCLIParamStorageContentTypeError() {
 func (suite *generateConfigTestSuite) TestCLIParamStorageBackgroundDownloadError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -741,10 +740,10 @@ func (suite *generateConfigTestSuite) TestCLIParamStorageBackgroundDownloadError
 func (suite *generateConfigTestSuite) TestCLIParamStorageInvalidateOnSyncNoError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -757,10 +756,10 @@ func (suite *generateConfigTestSuite) TestCLIParamStorageInvalidateOnSyncNoError
 func (suite *generateConfigTestSuite) TestCLIParamPreMountValidateNoError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -774,10 +773,10 @@ func (suite *generateConfigTestSuite) TestCLIParamPreMountValidateNoError() {
 func (suite *generateConfigTestSuite) TestInvalidLibfuseOption() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -794,10 +793,10 @@ func (suite *generateConfigTestSuite) TestInvalidLibfuseOption() {
 func (suite *generateConfigTestSuite) TestUndefinedLibfuseOption() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -814,10 +813,10 @@ func (suite *generateConfigTestSuite) TestUndefinedLibfuseOption() {
 func (suite *generateConfigTestSuite) TestInvalidUmaskValue() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -834,10 +833,10 @@ func (suite *generateConfigTestSuite) TestInvalidUmaskValue() {
 func (suite *generateConfigTestSuite) TestInvalidAttrTimeout() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -855,10 +854,10 @@ func (suite *generateConfigTestSuite) TestInvalidAttrTimeout() {
 func (suite *generateConfigTestSuite) TestInvalidEntryTimeout() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -876,10 +875,10 @@ func (suite *generateConfigTestSuite) TestInvalidEntryTimeout() {
 func (suite *generateConfigTestSuite) TestInvalidNegativeTimeout() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -896,7 +895,7 @@ func (suite *generateConfigTestSuite) TestInvalidNegativeTimeout() {
 func (suite *generateConfigTestSuite) TestEnvVarAccountName() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -911,7 +910,7 @@ func (suite *generateConfigTestSuite) TestEnvVarAccountName() {
 func (suite *generateConfigTestSuite) TestEnvVarAccountNameError() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())
@@ -925,10 +924,10 @@ func (suite *generateConfigTestSuite) TestEnvVarAccountNameError() {
 func (suite *generateConfigTestSuite) TestInvalidAccountType() {
 	defer suite.cleanupTest()
 	name := generateFileName()
-	v1ConfigFile, _ := ioutil.TempFile("", name+".tmp.cfg")
+	v1ConfigFile, _ := os.CreateTemp("", name+".tmp.cfg")
 	defer os.Remove(v1ConfigFile.Name())
 	v1ConfigFile.WriteString("accountName myAccountName\naccountType random")
-	v2ConfigFile, _ := ioutil.TempFile("", name+".tmp.yaml")
+	v2ConfigFile, _ := os.CreateTemp("", name+".tmp.yaml")
 	defer os.Remove(v2ConfigFile.Name())
 
 	outputFile := fmt.Sprintf("--output-file=%s", v2ConfigFile.Name())