+++ component/attr_cache/attr_cache_test.go
@@ -127,11 +127,15 @@ func assertUntouched(suite *attrCacheTestSuite, path string) {
 
 // Directory structure
 // a/
-//  a/c1/
-//   a/c1/gc1
-//	a/c2
+//
+//	 a/c1/
+//	  a/c1/gc1
+//		a/c2
+//
 // ab/
-//  ab/c1
+//
+//	ab/c1
+//
 // ac
 func generateNestedDirectory(path string) (*list.List, *list.List, *list.List) {
 	aPaths := list.New()