+++ component/azstorage/block_blob_test.go
@@ -1,3 +1,4 @@
+//go:build !authtest
 // +build !authtest
 
 /*
@@ -44,7 +45,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"path/filepath"
@@ -117,7 +117,7 @@ func uploadReaderAtToBlockBlob(ctx context.Context, reader io.ReaderAt, readerSi
 		if o.Progress != nil {
 			body = pipeline.NewRequestBodyProgress(body, o.Progress)
 		}
-		return blockBlobURL.Upload(ctx, body, o.BlobHTTPHeaders, o.Metadata, o.AccessConditions, o.BlobAccessTier, o.BlobTagsMap, o.ClientProvidedKeyOptions)
+		return blockBlobURL.Upload(ctx, body, o.BlobHTTPHeaders, o.Metadata, o.AccessConditions, o.BlobAccessTier, o.BlobTagsMap, o.ClientProvidedKeyOptions, azblob.ImmutabilityPolicyOptions{})
 	}
 
 	var numBlocks = uint16(((readerSize - 1) / o.BlockSize) + 1)
@@ -161,7 +161,7 @@ func uploadReaderAtToBlockBlob(ctx context.Context, reader io.ReaderAt, readerSi
 		return nil, err
 	}
 	// All put blocks were successful, call Put Block List to finalize the blob
-	return blockBlobURL.CommitBlockList(ctx, blockIDList, o.BlobHTTPHeaders, o.Metadata, o.AccessConditions, o.BlobAccessTier, o.BlobTagsMap, o.ClientProvidedKeyOptions)
+	return blockBlobURL.CommitBlockList(ctx, blockIDList, o.BlobHTTPHeaders, o.Metadata, o.AccessConditions, o.BlobAccessTier, o.BlobTagsMap, o.ClientProvidedKeyOptions, azblob.ImmutabilityPolicyOptions{})
 }
 
 type blockBlobTestSuite struct {
@@ -203,7 +203,7 @@ func (s *blockBlobTestSuite) SetupTest() {
 		os.Exit(1)
 	}
 
-	cfgData, _ := ioutil.ReadAll(cfgFile)
+	cfgData, _ := io.ReadAll(cfgFile)
 	err = json.Unmarshal(cfgData, &storageTestConfigurationParameters)
 	if err != nil {
 		fmt.Println("Failed to parse the config file")
@@ -402,11 +402,15 @@ func (s *blockBlobTestSuite) TestDeleteDir() {
 
 // Directory structure
 // a/
-//  a/c1/
-//   a/c1/gc1
-//	a/c2
+//
+//	 a/c1/
+//	  a/c1/gc1
+//		a/c2
+//
 // ab/
-//  ab/c1
+//
+//	ab/c1
+//
 // ac
 func generateNestedDirectory(path string) (*list.List, *list.List, *list.List) {
 	aPaths := list.New()
@@ -1208,7 +1212,7 @@ func (s *blockBlobTestSuite) TestWriteFile() {
 	file := s.containerUrl.NewBlobURL(name)
 	resp, err := file.Download(ctx, 0, int64(len(data)), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1230,7 +1234,7 @@ func (s *blockBlobTestSuite) TestTruncateSmallFileSmaller() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData[:truncatedLength], output)
 }
 
@@ -1256,7 +1260,7 @@ func (s *blockBlobTestSuite) TestTruncateChunkedFileSmaller() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData[:truncatedLength], output)
 }
 
@@ -1278,7 +1282,7 @@ func (s *blockBlobTestSuite) TestTruncateSmallFileEqual() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1304,7 +1308,7 @@ func (s *blockBlobTestSuite) TestTruncateChunkedFileEqual() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1326,7 +1330,7 @@ func (s *blockBlobTestSuite) TestTruncateSmallFileBigger() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output[:len(data)])
 }
 
@@ -1352,7 +1356,7 @@ func (s *blockBlobTestSuite) TestTruncateChunkedFileBigger() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output[:len(data)])
 }
 
@@ -1376,7 +1380,7 @@ func (s *blockBlobTestSuite) TestWriteSmallFile() {
 	dataLen := len(data)
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 
 	err = s.az.CopyToFile(internal.CopyToFileOptions{Name: name, File: f})
@@ -1401,7 +1405,7 @@ func (s *blockBlobTestSuite) TestOverwriteSmallFile() {
 	dataLen := len(data)
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 5, Data: newTestData})
@@ -1431,7 +1435,7 @@ func (s *blockBlobTestSuite) TestOverwriteAndAppendToSmallFile() {
 
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 5, Data: newTestData})
@@ -1462,7 +1466,7 @@ func (s *blockBlobTestSuite) TestAppendToSmallFile() {
 
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("-newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 9, Data: newTestData})
@@ -1493,7 +1497,7 @@ func (s *blockBlobTestSuite) TestAppendOffsetLargerThanSmallFile() {
 
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 12, Data: newTestData})
@@ -1528,7 +1532,7 @@ func (s *blockBlobTestSuite) TestAppendBlocksToSmallFile() {
 		BlockSize: 8,
 	})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("-newdata-newdata-newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 9, Data: newTestData})
@@ -1562,7 +1566,7 @@ func (s *blockBlobTestSuite) TestOverwriteBlocks() {
 		BlockSize: 4,
 	})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 16, Data: newTestData})
@@ -1596,7 +1600,7 @@ func (s *blockBlobTestSuite) TestOverwriteAndAppendBlocks() {
 		BlockSize: 4,
 	})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("43211234cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 32, Data: newTestData})
@@ -1629,7 +1633,7 @@ func (s *blockBlobTestSuite) TestAppendBlocks() {
 		BlockSize: 4,
 	})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("43211234cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: newTestData})
@@ -1662,7 +1666,7 @@ func (s *blockBlobTestSuite) TestAppendOffsetLargerThanSize() {
 		BlockSize: 4,
 	})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("43211234cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 45, Data: newTestData})
@@ -1686,7 +1690,7 @@ func (s *blockBlobTestSuite) TestCopyToFileError() {
 	defer s.cleanupTest()
 	// Setup
 	name := generateFileName()
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 
 	err := s.az.CopyToFile(internal.CopyToFileOptions{Name: name, File: f})
@@ -1701,7 +1705,7 @@ func (s *blockBlobTestSuite) TestCopyFromFile() {
 	testData := "test data"
 	data := []byte(testData)
 	homeDir, _ := os.UserHomeDir()
-	f, _ := ioutil.TempFile(homeDir, name+".tmp")
+	f, _ := os.CreateTemp(homeDir, name+".tmp")
 	defer os.Remove(f.Name())
 	f.Write(data)
 
@@ -1713,7 +1717,7 @@ func (s *blockBlobTestSuite) TestCopyFromFile() {
 	file := s.containerUrl.NewBlobURL(name)
 	resp, err := file.Download(ctx, 0, int64(len(data)), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
-	output, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1737,7 +1741,7 @@ func (s *blockBlobTestSuite) TestCreateLink() {
 	s.assert.EqualValues("true", props.NewMetadata()[symlinkKey])
 	resp, err := link.Download(ctx, 0, props.ContentLength(), azblob.BlobAccessConditions{}, false, azblob.ClientProvidedKeyOptions{})
 	s.assert.Nil(err)
-	data, _ := ioutil.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
+	data, _ := io.ReadAll(resp.Body(azblob.RetryReaderOptions{}))
 	s.assert.EqualValues(target, data)
 }