+++ component/azstorage/datalake_test.go
@@ -1,3 +1,4 @@
+//go:build !authtest
 // +build !authtest
 
 /*
@@ -41,7 +42,7 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"math/rand"
 	"os"
 	"strings"
@@ -91,7 +92,7 @@ func (s *datalakeTestSuite) SetupTest() {
 		os.Exit(1)
 	}
 
-	cfgData, _ := ioutil.ReadAll(cfgFile)
+	cfgData, _ := io.ReadAll(cfgFile)
 	err = json.Unmarshal(cfgData, &storageTestConfigurationParameters)
 	if err != nil {
 		fmt.Println("Failed to parse the config file")
@@ -726,7 +727,7 @@ func (s *datalakeTestSuite) TestWriteSmallFile() {
 	dataLen := len(data)
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 
 	err = s.az.CopyToFile(internal.CopyToFileOptions{Name: name, File: f})
@@ -751,7 +752,7 @@ func (s *datalakeTestSuite) TestOverwriteSmallFile() {
 	dataLen := len(data)
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 5, Data: newTestData})
@@ -781,7 +782,7 @@ func (s *datalakeTestSuite) TestOverwriteAndAppendToSmallFile() {
 
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 5, Data: newTestData})
@@ -812,7 +813,7 @@ func (s *datalakeTestSuite) TestAppendOffsetLargerThanSmallFile() {
 
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 12, Data: newTestData})
@@ -843,7 +844,7 @@ func (s *datalakeTestSuite) TestAppendToSmallFile() {
 
 	_, err := s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("-newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 9, Data: newTestData})
@@ -883,7 +884,7 @@ func (s *datalakeTestSuite) TestAppendBlocksToSmallFile() {
 			BlockSize: 8,
 		})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("-newdata-newdata-newdata")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 9, Data: newTestData})
@@ -923,7 +924,7 @@ func (s *datalakeTestSuite) TestOverwriteBlocks() {
 			BlockSize: 4,
 		})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 16, Data: newTestData})
@@ -963,7 +964,7 @@ func (s *datalakeTestSuite) TestOverwriteAndAppendBlocks() {
 			BlockSize: 4,
 		})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("43211234cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 32, Data: newTestData})
@@ -1001,7 +1002,7 @@ func (s *datalakeTestSuite) TestAppendBlocks() {
 			BlockSize: 4,
 		})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("43211234cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: newTestData})
@@ -1039,7 +1040,7 @@ func (s *datalakeTestSuite) TestAppendOffsetLargerThanSize() {
 			BlockSize: 4,
 		})
 	s.assert.Nil(err)
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 	newTestData := []byte("43211234cake")
 	_, err = s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 45, Data: newTestData})
@@ -1329,7 +1330,7 @@ func (s *datalakeTestSuite) TestWriteFile() {
 	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
 	resp, err := file.Download(ctx, 0, int64(len(data)))
 	s.assert.Nil(err)
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1351,7 +1352,7 @@ func (s *datalakeTestSuite) TestTruncateSmallFileSmaller() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength))
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData[:truncatedLength], output)
 }
 
@@ -1378,7 +1379,7 @@ func (s *datalakeTestSuite) TestTruncateChunkedFileSmaller() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength))
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData[:truncatedLength], output)
 }
 
@@ -1400,7 +1401,7 @@ func (s *datalakeTestSuite) TestTruncateSmallFileEqual() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength))
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1427,7 +1428,7 @@ func (s *datalakeTestSuite) TestTruncateChunkedFileEqual() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength))
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1449,7 +1450,7 @@ func (s *datalakeTestSuite) TestTruncateSmallFileBigger() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength))
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output[:len(data)])
 }
 
@@ -1476,7 +1477,7 @@ func (s *datalakeTestSuite) TestTruncateChunkedFileBigger() {
 	resp, err := file.Download(ctx, 0, int64(truncatedLength))
 	s.assert.Nil(err)
 	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output[:len(data)])
 }
 
@@ -1499,7 +1500,7 @@ func (s *datalakeTestSuite) TestCopyToFile() {
 	data := []byte(testData)
 	dataLen := len(data)
 	s.az.WriteFile(internal.WriteFileOptions{Handle: h, Offset: 0, Data: data})
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 
 	err := s.az.CopyToFile(internal.CopyToFileOptions{Name: name, File: f})
@@ -1518,7 +1519,7 @@ func (s *datalakeTestSuite) TestCopyToFileError() {
 	defer s.cleanupTest()
 	// Setup
 	name := generateFileName()
-	f, _ := ioutil.TempFile("", name+".tmp")
+	f, _ := os.CreateTemp("", name+".tmp")
 	defer os.Remove(f.Name())
 
 	err := s.az.CopyToFile(internal.CopyToFileOptions{Name: name, File: f})
@@ -1533,7 +1534,7 @@ func (s *datalakeTestSuite) TestCopyFromFile() {
 	testData := "test data"
 	data := []byte(testData)
 	homeDir, _ := os.UserHomeDir()
-	f, _ := ioutil.TempFile(homeDir, name+".tmp")
+	f, _ := os.CreateTemp(homeDir, name+".tmp")
 	defer os.Remove(f.Name())
 	f.Write(data)
 
@@ -1545,7 +1546,7 @@ func (s *datalakeTestSuite) TestCopyFromFile() {
 	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
 	resp, err := file.Download(ctx, 0, int64(len(data)))
 	s.assert.Nil(err)
-	output, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1570,7 +1571,7 @@ func (s *datalakeTestSuite) TestCreateLink() {
 	s.assert.EqualValues("true", metadata["Is_symlink"])
 	resp, err := link.Download(ctx, 0, props.ContentLength())
 	s.assert.Nil(err)
-	data, _ := ioutil.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	data, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
 	s.assert.EqualValues(target, data)
 }