+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 2.0.4 (WIP)
 **Features**
-- Added new config parameter 'refresh-sec' in 'file-cache'. When file-cache-timeout is set to a large value, this field can control when to refresh the file if container has a file with higher LMT compared to one in the local cache.
+- Added new config parameter "max-fuse-threads" under "libfuse" config to control max threads allowed at libfuse layer.
+- Added new config parameter 'refresh-sec' in 'file-cache'. When file-cache-timeout is set to a large value, this field can control when to refresh the file if file in container has changed.
 
+**Bug Fixes**
 
 ## 2.0.3 (2023-04-26)
 **Bug Fixes**